weibo/requirements.txt
Package                Version
---------------------- -------
adium-theme-ubuntu     0.3.4
alembic                1.4.2
click                  7.1.2
dnspython              1.16.0
dominate               2.5.1
ecdsa                  0.13
email-validator        1.1.0
Flask                  1.1.2
Flask-Bootstrap        3.3.7.1
Flask-CKEditor         0.4.4.1
Flask-Login            0.5.0
Flask-Migrate          2.5.3
Flask-Moment           0.10.0
Flask-Script           2.0.6
Flask-SQLAlchemy       2.4.1
Flask-WTF              0.14.3
gyp                    0.1
idna                   2.9
itsdangerous           1.1.0
Jinja2                 2.11.2
Mako                   1.1.2
MarkupSafe             1.1.1
mysql-connector-python 2.0.4
mysql-utilities        1.6.1
paramiko               1.16.0
pexpect                4.0.1
Pillow                 6.2.2
pip                    20.1
pkg-resources          0.0.0
ptyprocess             0.5
pycrypto               2.6.1
pygobject              3.20.0
PyMySQL                0.9.3
pyodbc                 3.0.7
pysqlite               2.7.0
python-dateutil        2.8.1
python-dotenv          0.12.0
python-editor          1.0.4
setuptools             45.0.0
six                    1.14.0
SQLAlchemy             1.3.17
typing                 3.7.4.3
unity-lens-photos      1.0
vboxapi                1.0
virtualenv             15.0.1
visitor                0.1.3
Werkzeug               1.0.1
wheel                  0.34.2
WTForms                2.3.1
